:recycle: Rename api/delegate to api/delegates
