+++ src/actions/search.test.js
@@ -3,7 +3,7 @@ import networks from '../constants/networks';
 import txFilters from '../constants/transactionFilters';
 import * as searchAPI from '../utils/api/search';
 import * as accountAPI from '../utils/api/account';
-import * as delegateAPI from '../utils/api/delegate';
+import * as delegateAPI from '../utils/api/delegates';
 import * as transactionsAPI from '../utils/api/transactions';
 import accounts from '../../test/constants/accounts';
 import * as actions from './search';
@@ -18,7 +18,7 @@ const {
 
 jest.mock('../utils/api/search');
 jest.mock('../utils/api/account');
-jest.mock('../utils/api/delegate');
+jest.mock('../utils/api/delegates');
 jest.mock('../utils/api/transactions');
 
 describe('actions: search', () => {
